Move unit tests for moved Game methods to a new spec/core_spec.py file
